correct labeling of main file
