+++ IDEA/main.py
@@ -21,44 +21,6 @@ from lookups import *
 from web_classes import *
 from web_methods import *
 
-# from dotenv import load_dotenv
-
-# load_dotenv()
-
-
-############## DATABASE
-
-# Establish connection to server
-
-client = MongoClient(DB_URI,server_api=ServerApi('1'))
-
-# Ping server on startup
-
-try:
-    client.admin.command('ping')
-    print("Connection Successful")
-except Exception as e:
-    print(e)
-
-# Method to get data of server
-
-@st.cache_data(ttl=1200)
-def get_data():
-    db=client["AIME"]
-    items=db["Conversation"].find()
-    items = list(items)
-    return items
-
-# MongoDB Collection to add to
-
-collection=client["AIME"]["Conversation"]
-
-# Collection for surveys for FIRST BETATEST
-
-survey = {}
-
-######### WEBSITE #########
-
 st.set_page_config(page_title = "AIME",
                    page_icon = "🧑‍⚕️",
                    layout = "wide",
@@ -69,7 +31,6 @@ if "stage" not in st.session_state:
 
 def set_stage(stage):
     st.session_state["stage"] = stage
-
 
 if st.session_state["stage"] == LOGIN_PAGE:
     st.write("Welcome, and thank you for volunteering to participate in this beta test! This is an application where you will virtually simulate an interview with a patient, provide a differential diagnosis for them, and then automatically receive grading and feedback based on your performance.")
@@ -90,37 +51,47 @@ if st.session_state["stage"] == LOGIN_PAGE:
                 st.session_state["username"] = username
                 st.write("Authentication successful!")
                 time.sleep(1)
-                set_stage(CHAT_SETUP)
+                set_stage(SETTINGS)
                 st.rerun()
             else:
                 st.write("Password incorect.")
 
 
-if st.session_state["stage"] == CHAT_SETUP:
+if st.session_state["stage"] == SETTINGS:
     st.session_state["interview"] = None
     st.session_state["messages"] = []
     st.session_state["convo_memory"] = []
     st.session_state["convo_summary"]=""
     st.session_state["convo_file"] = None
     st.session_state["convo_prompt"] = ""
-    st.session_state["sent"] = False
+    # st.session_state["sent"] = False
     st.session_state["start_time"] = date.datetime.now()
     st.session_state["tokens"] = {"convo": {"input": 0, "output": 0},
                                   "class": {"input": 0, "output": 0},
                                   "diag": {"input": 0, "output": 0}}
 
-    patient_name = "Jeffrey Smith"
+    layout1 = st.columns([1, 3, 1])
+    with layout1[1]:
+        st.title("Patient Settings")
+        patient_name = st.selectbox("Which patient would you like to interview?", 
+                                    ["Jeffrey Smith"],
+                                    index = None,
+                                    placeholder = "Select patient...")
+        if patient_name: 
+            st.session_state["interview"] = Interview.build(username = st.session_state["username"], 
+                                                            patient = Patient.build(patient_name))
 
-    st.session_state["interview"] = Interview.build(username = st.session_state["username"], 
-                                                    patient = Patient.build(patient_name))
+        if st.session_state["interview"]: 
+            # st.session_state["sent"]==False
+            st.button("Start Interview", on_click=set_stage, args=[CHAT_SETUP])
 
-    st.session_state["sent"]==False
 
+if st.session_state["stage"] == CHAT_SETUP:
     st.session_state["convo_prompt"] = st.session_state["interview"].patient.convo_prompt
-    if(st.session_state["sent"]==False):
-        st.session_state["interview"].start_time = str(st.session_state["start_time"])
-        collection.insert_one(st.session_state["interview"].model_dump())
-        st.session_state["sent"]==True
+    # if(st.session_state["sent"]==False):
+    #     st.session_state["interview"].start_time = str(st.session_state["start_time"])
+        # collection.insert_one(st.session_state["interview"].model_dump())
+        # st.session_state["sent"]==True
 
     set_stage(CHAT_INTERFACE_VOICE)
 
@@ -162,70 +133,52 @@ if st.session_state["stage"] == CHAT_INTERFACE_VOICE:
 
 
 if st.session_state["stage"] == KEY_PHYSICALS:
-    st.title("Key Physical Exam Information")
+    layout1 = st.columns([2, 3, 2])
+    with layout1[1]:
+        st.title("Key Physical Exam Information")
+        findings = st.text_area(label = "Based on your interview with the patient, state the key physical examination findings that you would look for. You will be provided the actual exam results on the next screen.",
+                                height = 200)
 
-    st.button("Next", on_click=set_stage, args=[PHYSICAL_ECG_SCREEN])
+        layout11 = st.columns([1, 1, 1])
+        if layout11[1].button("Next", use_container_width=True):
+            st.session_state["interview"].add_key_findings(findings)
+            set_stage(PHYSICAL_ECG_SCREEN)
+            st.rerun()
 
 
 if st.session_state["stage"] == PHYSICAL_ECG_SCREEN:
-    st.title("Physical and ECG Information")
-    layout1 = st.columns([7, 1])
-    layout1[0].write("Now that you've taken a chance to speak with the patient, you can take a chance to take a look at data obtained during a physical upon admittance to the ER. Review the following physical and ECG before proceeding.")
-    layout1[1].button("Proceed to Diagnosis", on_click=set_stage, args = [DIAGNOSIS])
     
-    # Update the database from before
-
-    collection.replace_one({"username":st.session_state["interview"].username, "start_time":st.session_state["interview"].start_time}, st.session_state["interview"].model_dump())
-
-    layout2 = st.columns([1, 1])
-    with layout2[0].container():
-        st.header("Physical Examination", divider = "grey")
+    layout1 = st.columns([1, 3, 1])
+    with layout1[1].container():
+        st.title("Physical Examination")
+        st.write("Now that you've taken a chance to speak with the patient, you can take a chance to take a look at data obtained during a physical upon admittance to the ER. Review it before proceeding.")
+        
+        st.divider()
         physical_exam_doc = Document(st.session_state["interview"].patient.physical)
         for paragraph in physical_exam_doc.paragraphs:
             st.write(paragraph.text)
+        st.divider()
+
+        layout11 = st.columns([1, 1, 1])
+        layout11[1].button("Proceed to Diagnosis", on_click=set_stage, args = [DIAGNOSIS], use_container_width=True)
 
 
 if st.session_state["stage"] == DIAGNOSIS:
     st.title("Diagnosis")
     st.write("Use the interview transcription and additional patient information (physical examination and ECG) to provide an interpretative summary, a list of potential diagnoses, a rationale reasoning through which diagnoses are more/less likely, and a final diagnosis.")
     st.write("Click the \"Get Feedback\" button once you are all done to automatically receive grading and feedback on your interview and diagnosis.")
+    st.divider()
 
     # 2 column full width layout
     layout1 = st.columns([1, 1])
 
     # User inputs
     hpi = layout1[0].text_area(label = "History of Present Illness", height = 200)
-    past_histories = layout1[0].text_area(label = "Past histories (include relevant past, family, and social histories and review of symptoms)", height = 200)
+    past_histories = layout1[0].text_area(label = "Past Histories (include relevant past, family, and social histories and review of symptoms)", height = 200)
     summary = layout1[0].text_area(label = "Summary Statement", height = 200)
-    assessment = layout1[0].text_area(label = "Assessment - specify most likely diagnosis and alternative diagnoses. Include your reasoning/explanation for each differential diagnoses.", height = 200)
-    plan = layout1[0].text_area(label = "Plan - include your reasoning for your initial diagnostic plan.", height = 200)
+    assessment = layout1[0].text_area(label = "Assessment - Specify most likely diagnosis and alternative diagnoses. Include your reasoning/explanation for each differential diagnosis.", height = 200)
+    plan = layout1[0].text_area(label = "Plan - State your initial diagnostic plan. Includ eyour reasoning/explanation for the plan.", height = 200)
 
-    # 3 buttons at bottom
-    layout12 = layout1[0].columns([1, 1, 1])
-    # New Interview
-    # layout12[0].button("New Interview", on_click=set_stage, args=[SETTINGS])
-    # Download Interview
-    # currentDateAndTime = date.datetime.now()
-    # st.session_state["interview"].date_time = str(currentDateAndTime)
-    # date_time = currentDateAndTime.strftime("%d-%m-%y__%H-%M")
-    
-    bio = io.BytesIO()
-    st.session_state["convo_file"] = create_convo_file(st.session_state["interview"].username, 
-                                                       st.session_state["interview"].patient.id, 
-                                                       [message.model_dump() for message in st.session_state["interview"].messages])
-    st.session_state["convo_file"].save(bio)
-
-    date_time = date.datetime.now().strftime("%d-%m-%y__%H-%M")
-    layout12[1].download_button("Download interview", 
-                                data = bio.getvalue(), 
-                                file_name = st.session_state["interview"].username + "_"+date_time + ".docx", 
-                                mime = "docx")
-    # Get Feedback
-    if layout12[2].button("Get Feedback"): 
-        st.session_state["interview"].add_diagnosis_inputs(summary, [potential1, potential2, potential3], rationale, final)
-        set_stage(FEEDBACK_SETUP)
-        st.rerun()
-    
     # Interview transcription
     chat_container = layout1[1].container(height=400)
     for message in st.session_state["interview"].messages:
@@ -237,15 +190,71 @@ if st.session_state["stage"] == DIAGNOSIS:
         physical_exam_doc = Document(st.session_state["interview"].patient.physical)
         for paragraph in physical_exam_doc.paragraphs:
             st.write(paragraph.text)
-    # ECG
-    with layout1[1].expander("ECG"):
-        st.image(st.session_state["interview"].patient.ECG)
 
+    # 3 buttons: New Interview, Download Interview, Get Feedback
+    st.divider()
+    layout2 = st.columns([1, 1, 1, 1, 1])
+
+    # New Interview
+    layout2[1].button("New Interview", on_click=set_stage, args=[SETTINGS], use_container_width=True)
     
+    # # Download Interview
+    # currentDateAndTime = date.datetime.now()
+    # st.session_state["interview"].date_time = str(currentDateAndTime)
+    # date_time = currentDateAndTime.strftime("%d-%m-%y__%H-%M")
+
+    # bio = io.BytesIO()
+    # st.session_state["convo_file"] = create_convo_file(st.session_state["interview"].username, 
+    #                                                    st.session_state["interview"].patient.id, 
+    #                                                    [message.model_dump() for message in st.session_state["interview"].messages])
+    # st.session_state["convo_file"].save(bio)
+
+    # date_time = date.datetime.now().strftime("%d-%m-%y__%H-%M")
+    # layout12[1].download_button("Download interview", 
+    #                             data = bio.getvalue(), 
+    #                             file_name = st.session_state["interview"].username + "_" + date_time + ".docx", 
+    #                             mime = "docx")
+    
+    # Test cases
+    layout21 = layout2[2].columns([1, 1])
+    if layout21[0].button("TEST: BAD", use_container_width=True):
+        with open("./IDEA/test_cases/bad.json", "r", encoding="utf8") as bad_json:
+            bad_case = json.load(bad_json)
+            # print(bad_case)
+            # print("\n\n")
+            st.session_state["interview"].add_key_findings(bad_case["Key Findings"])
+            st.session_state["interview"].add_other_inputs(bad_case["HPI"], 
+                                                           bad_case["Past Histories"], 
+                                                           bad_case["Summary"], 
+                                                           bad_case["Assessment"], 
+                                                           bad_case["Plan"])
+            # print(st.session_state["interview"].post_note_inputs)
+            # print("\n\n")
+        set_stage(FEEDBACK_SETUP)
+        st.rerun()
+    if layout21[1].button("TEST: GOOD", use_container_width=True):
+        with open("./IDEA/test_cases/good.json", "r", encoding="utf8") as good_json:
+            good_case = json.load(good_json)
+            st.session_state["interview"].add_key_findings(good_case["Key Findings"])
+            st.session_state["interview"].add_other_inputs(good_case["HPI"], 
+                                                           good_case["Past Histories"], 
+                                                           good_case["Summary"], 
+                                                           good_case["Assessment"], 
+                                                           good_case["Plan"])
+        set_stage(FEEDBACK_SETUP)
+        st.rerun()
+    
+    # Get Feedback
+    if layout2[3].button("Get Feedback", use_container_width=True): 
+        st.session_state["interview"].add_post_note(hpi, past_histories, summary, assessment, plan)
+        set_stage(FEEDBACK_SETUP)
+        st.rerun()
+
+
 if st.session_state["stage"] == FEEDBACK_SETUP:
      # Update the database from before
 
-    collection.replace_one({"username":st.session_state["interview"].username, "start_time":st.session_state["interview"].start_time}, st.session_state["interview"].model_dump())
+    # collection.replace_one({"username":st.session_state["interview"].username, "start_time":st.session_state["interview"].start_time}, st.session_state["interview"].model_dump())
 
     st.title("Processing feedback...")
     st.write("This might take a few minutes.")
@@ -260,100 +269,7 @@ if st.session_state["stage"] == FEEDBACK_SETUP:
 if st.session_state["stage"] == FEEDBACK_SCREEN:
     st.title("Feedback")
     layout1 = st.columns([7, 1])
-    layout1[0].write("This is the WIP (the UI especially) feedback screen. There are 3 tabs of grading/feedback available. \"Data Acquisition\" looks through your interview with the patient and grades you on what information you asked for and what information you obtained from the patient. \"Diagnosis\" is self explanatory, grading you on your summary, diagnoses, and rationale. \"Case Explanation\" provides you with a PDF of a detailed explanation of the John Smith case written by Dr. Corbett.")
-    layout1[0].write("Click the \"Go to Survey\" button on the right once you are done looking through the grading/feedback; the next (final) screen will take you through a brief survey on your experience with the application")
-    layout1[1].button("Go to Survey", on_click=set_stage, args=[SURVEY])
+    layout1[0].write("This is the WIP feedback screen.")
     
     # Let the display methods cook
     display_Interview(st.session_state["interview"].model_dump())
-
-
-if st.session_state["stage"] == SURVEY:
-
-     # Update the database from before
-
-    collection.replace_one({"username":st.session_state["interview"].username, "start_time":st.session_state["interview"].start_time}, st.session_state["interview"].model_dump())
-
-    layout1 = st.columns([1, 3, 1])
-    with layout1[1]:
-        st.title("Survey")
-        st.write("Please take the time to give us some feedback and constructive criticism so that we can improve this application.")
-
-        #QUESTION 1
-        question1="Were there any issues you encountered while interviewing the virtual patient or writing your diagnosis? What could be improved?"
-        answer1 = st.text_area(label = question1, height = 100)
-        #QUESTION 2
-        question2="Were there any issues you encountered in your grading and feedback? What could be improved?"
-        answer2 = st.text_area(label = question2, height = 100)
-        #QUESTION 3
-        question3="Do you have any other suggestions or areas for improvement? This could include ideas for new features, feedback sections, etc."
-        answer3 = st.text_area(label = question3, height = 100)
-        
-        question4="How helpful do you think practicing an interview would be in learning cardiovascular pathophisiology?"
-        answer4= st.text_area(label = question4, height = 100)
-
-
-        #STORING INTERVIEWS IN SEPARATE QUESTIONS DB
-        survey["question1"]=[question1,answer1]
-        survey["question2"]=[question2,answer2]
-        survey["question3"]=[question3,answer3]
-        survey["question4"]=[question4,answer4]
-
-        st.session_state["interview"].survey = survey
-        columns = st.columns(3)
-        columns[1].button("Go to End Screen", on_click=set_stage, args=[FINAL_SCREEN])
-
-
-if st.session_state["stage"] == FINAL_SCREEN: 
-    layout1 = st.columns([1, 3, 1])
-    with layout1[1]:
-        st.title("Thank you! :heart:")
-        st.write("All done! Thank you so much for taking the time to help us test our application. Your interview, diagnosis, and survey has been recorded and sent to us automatically.")
-        st.write("Click the download button to download your most recent interview as a word file. Click the New Interview button to go back to the chat interface and keep testing.")
-        
-        # Record time
-        end_time = date.datetime.now()
-        time_elapsed = end_time - st.session_state["start_time"]
-        st.session_state["interview"].time_elapsed = str(time_elapsed)
-
-        # Record cost
-        print(str(st.session_state["tokens"]) + "\n")
-        st.session_state["interview"].cost = 0
-        st.session_state["interview"].cost += (st.session_state["tokens"]["convo"]["input"] / 1000000) * COSTS[CONVO_MODEL]["input"]
-        st.session_state["interview"].cost += (st.session_state["tokens"]["convo"]["output"] / 1000000) * COSTS[CONVO_MODEL]["output"]
-        st.session_state["interview"].cost += (st.session_state["tokens"]["class"]["input"] / 1000000) * COSTS[CLASS_MODEL]["input"]
-        st.session_state["interview"].cost += (st.session_state["tokens"]["class"]["output"] / 1000000) * COSTS[CLASS_MODEL]["output"]
-        st.session_state["interview"].cost += (st.session_state["tokens"]["diag"]["input"] / 1000000) * COSTS[DIAG_MODEL]["input"]
-        st.session_state["interview"].cost += (st.session_state["tokens"]["diag"]["output"] / 1000000) * COSTS[DIAG_MODEL]["output"]
-        st.write("Estimated cost: $" + str(st.session_state["interview"].cost))
-
-
-        bio = io.BytesIO()
-        st.session_state["convo_file"] = create_convo_file(st.session_state["interview"].username, 
-                                                        st.session_state["interview"].patient.id, 
-                                                        [message.model_dump() for message in st.session_state["interview"].messages])
-        st.session_state["convo_file"].save(bio)
-        
-        date_time = end_time.strftime("%d-%m-%y__%H-%M")
-        button_columns = st.columns(3)
-        button_columns[0].download_button("Download interview", 
-                        data = bio.getvalue(),
-                        file_name = st.session_state["interview"].username + "_"+ date_time + ".docx",
-                        mime = "docx")  
-         # Update the database from before
-
-        collection.replace_one({"username":st.session_state["interview"].username, "start_time":st.session_state["interview"].start_time}, st.session_state["interview"].model_dump())
-
-        # # Store interview in database and send email as backup
-        # if st.session_state["sent"] == False:
-        #     collection.insert_one(st.session_state["interview"].model_dump())
-        #     # send_email(bio, EMAIL_TO_SEND, st.session_state["interview"].username, date_time, None)
-        #     st.session_state["sent"] = True
-            
-        # st.download_button("Download JSON",
-        #             data=st.session_state["interview"].get_json(),
-        #             file_name = st.session_state ["interview"].get_username() + "_"+date_time + ".json",
-        #             mime="json")
-
-        button_columns[1].button("New Interview", on_click=set_stage, args=[SETTINGS])
-        button_columns[2].button("Back to Login", on_click=set_stage, args=[LOGIN_PAGE])
